requirements.txt
PyQt5==5.15.9
bcrypt==4.1.2
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2
xlrd>=2.0.1
Pillow==10.1.0
matplotlib==3.8.2
schedule==1.2.0
qrcode[pil]>=7.4.2
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
pymysql>=1.1.0

